requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
anyio==4.12.1
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bitsandbytes==0.49.2
bleach==6.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.8.3
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.2
idna==3.11
importlib_metadata==7.2.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiwer==3.0.0
json5==0.13.0
jsonpointer==3.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.5.0
lark==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==1.26.4
packaging==23.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
peft==0.18.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.9.4
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
rapidfuzz==2.13.7
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.30.0
safetensors==0.7.0
Send2Trash==2.1.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
stack-data==0.6.3
streamlit==1.30.0
sympy==1.14.0
tenacity==8.5.0
terminado==0.18.1
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
trl==0.29.1
typer==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.6.3
validators==0.35.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
